Refactor test provisioning to dialect-level files
Fixes: #5085
