+++ lib/sqlalchemy/dialects/postgresql/__init__.py
@@ -7,6 +7,7 @@
 
 from . import base
 from . import pg8000  # noqa
+from . import provision  # noqa
 from . import psycopg2  # noqa
 from . import psycopg2cffi  # noqa
 from . import pygresql  # noqa
